requirements.txt
allure-pytest==2.13.0
allure-python-commons==2.13.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
h11==0.14.0
identify==2.5.19
idna==3.4
iniconfig==2.0.0
nodeenv==1.7.0
outcome==1.2.0
packaging==23.0
platformdirs==3.1.0
pluggy==1.0.0
pre-commit==3.1.1
PySocks==1.7.1
pytest==7.2.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
virtualenv==20.20.0
webdriver-manager==3.8.5
wsproto==1.2.0
